+++ composer.json
@@ -6,6 +6,7 @@
     "keywords": [
         "sonata",
         "formatter",
+        "editor",
         "ckeditor"
     ],
     "authors": [
