+++ src/Admin/CkeditorAdminExtension.php
@@ -16,6 +16,7 @@ namespace Sonata\FormatterBundle\Admin;
 use Sonata\AdminBundle\Admin\AbstractAdminExtension;
 use Sonata\AdminBundle\Admin\AdminInterface;
 use Sonata\AdminBundle\Route\RouteCollectionInterface;
+use Sonata\FormatterBundle\Controller\CkeditorAdminController;
 
 /**
  * Adds browser and upload routes to the Admin.
@@ -29,11 +30,11 @@ final class CkeditorAdminExtension extends AbstractAdminExtension
     public function configureRoutes(AdminInterface $admin, RouteCollectionInterface $collection): void
     {
         $collection->add('ckeditor_browser', 'ckeditor_browser', [
-            '_controller' => 'SonataFormatterBundle:CkeditorAdmin:browser',
+            '_controller' => CkeditorAdminController::class.'::browserAction',
         ]);
 
         $collection->add('ckeditor_upload', 'ckeditor_upload', [
-            '_controller' => 'SonataFormatterBundle:CkeditorAdmin:upload',
+            '_controller' => CkeditorAdminController::class.'::uploadAction',
         ]);
     }
 }
